Fix migration columns

+++ database/migrations/2020_12_11_170000_create_polis_table.php
@@ -16,7 +16,7 @@ class CreatePolisTable extends Migration
         Schema::create('polis', function (Blueprint $table) {
             $table->id();
             $table->string('nama');
-            $table->integer('nomor_sekarang')->nullable();
+            $table->unsignedSmallInteger('nomor_sekarang')->nullable();
             $table->timestamps();
         });
     }
